extract stuff to util
